Use pre_trained_model? If yes....
Hi,

We currently have in section 3.3. (Using weights from a pre-trained model as initial weights):
"Use pre_trained_model? If *not*, please provide the path to the model folder:..."
My understanding is that it should say 
"Use pre_trained_model? If *yes*, please provide the path to the model folder:..."

Thanks