Memory reduced and input shapes fixed
The RAM memory that the notebook has been reduced, even so the training time has increase a little.

The problem with the inputs of different sizes (which before was not able) has been fixed. Now, instead 
of splitting the input image in a number of patches, a patch size is selected and random patches are
extracted from the input image. Allowing with this to the user to choose the patch size and fixing
any possible problem with the input size.