Update EmbedSeg_2D_ZeroCostDL4Mic.ipynb
Uninstalling the incompatible decencies and re-installing tested versions that met the dependency requirement solved PyPI related issues. 

After reviewing the EmbedSeg PyPI release history and comparing its source code with the EmbedSeg GitHub repo, I noticed a number of changes, of which all resulted in our version of notebook to not run properly. 

So in the section relating to installing key dependencies, I did change the path that PIP retrieves EmbedSeg. From here on forward, we are installing EmbedSeg from its GitHub repo (which is by far the most updated version) 

%pip install git+https://github.com/juglab/EmbedSeg.git

Another point to consider was a minor change in the original train section of EmbedSeg. The variable normalization_factor in train_dataset_dict and val_dataset_dict are no longer needed on the updated EmbedSeg version, hence the two are muted. 

After implementing these changes and running the notebook, I did not encounter any issues.